encrypt_decrypt/requirements.txt
Flask
pycryptodome
qrcode
Pillow
pyzbar
